The derivation of a linear-time algorithm.
Using this derivation, we can calculate, in linear time, the the slope of the line of best fit for every possible window of a fixed size.  Our current algorithm chops the calculation up into pieces, but therefore has the potential to miss forms of performance degradation that happen between window boundaries.  For example, if performance suddenly gets worse at midnight December 31st, and the new year begins with depressed performance, our current algorithm will likely not fire an alert.

The new algorithm (unimplemented, but derived in the contained files) will not miss this problem, because it will try all year-long windows, not just the data windows that start on Jan 1 and end on Dec 31.